modify log component to accept external style
